requirements.txt
asgiref==3.7.2
astroid==2.6.6
atomicwrites==1.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.11.0
certifi==2020.12.5
chardet==4.0.0
click==8.1.7
colorama==0.4.6
coverage==7.3.2
Django==4.2.8
django-ckeditor==6.7.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
dnspython==2.4.2
email-validator==1.1.2
Faker==25.8.0
idna==2.10
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.6.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
profanityfilter==2.0.6
py==1.11.0
pylint==2.7.4
pytest==6.2.5
pytest-django==4.8.0
python-dotenv==0.17.0
pytz==2023.3.post1
requests==2.25.1
sentinels==1.0.0
soupsieve==2.2.1
sqlparse==0.4.4
stripe==7.4.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
urllib3==1.26.5
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
